whitea/requirements.txt
PyNaCl==1.5.0
beautifulsoup4==4.11.1
discord.py==1.7.3
requests==2.28.0
